# hub/frontend/openapi/wardn-hub-api.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegistryListMetadata": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "nextCursor": {
            "default": "",
            "title": "Nextcursor",
            "type": "string"
          }
        },
        "required": [
          "count"
        ],
        "title": "RegistryListMetadata",
        "type": "object"
      },
      "RegistryServerDetailResponse": {
        "properties": {
          "server": {
            "$ref": "#/components/schemas/RegistryServerRead"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/RegistryServerVersionRead"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "server"
        ],
        "title": "RegistryServerDetailResponse",
        "type": "object"
      },
      "RegistryServerListResponse": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/RegistryListMetadata"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/RegistryServerRead"
            },
            "title": "Servers",
            "type": "array"
          }
        },
        "required": [
          "servers",
          "metadata"
        ],
        "title": "RegistryServerListResponse",
        "type": "object"
      },
...
      "RegistryServerVersionCreate": {
        "additionalProperties": true,
        "properties": {
          "$schema": {
            "examples": [
              "https://static.modelcontextprotocol.io/schemas/2025-12-11/server.schema.json"
            ],
            "minLength": 1,
            "title": "$Schema",
            "type": "string"
          },
          "_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "icons": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Icons",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$",
            "title": "Name",
            "type": "string"
          },
          "packages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Packages",
            "type": "array"
          },
          "remotes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Remotes",
            "type": "array"
          },
          "repository": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository"
          },
          "title": {
            "default": "",
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          },
          "websiteUrl": {
            "default": "",
            "maxLength": 2048,
            "title": "Websiteurl",
            "type": "string"
          }
        },
        "required": [
          "$schema",
          "name",
          "description",
          "version"
        ],
        "title": "RegistryServerVersionCreate",
        "type": "object"
      },
      "RegistryServerVersionDetailResponse": {
        "properties": {
          "approval": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval"
          },
          "server": {
            "$ref": "#/components/schemas/RegistryServerRead"
          },
          "support": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support"
          },
          "version": {
            "$ref": "#/components/schemas/RegistryServerVersionRead"
          }
        },
        "required": [
          "server",
          "version"
        ],
        "title": "RegistryServerVersionDetailResponse",
        "type": "object"
      },
      "RegistryServerVersionListResponse": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/RegistryListMetadata"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/RegistryServerVersionRead"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "versions",
          "metadata"
        ],
        "title": "RegistryServerVersionListResponse",
        "type": "object"
      },
...
      "RegistryServerVersionUpdate": {
        "additionalProperties": true,
        "properties": {
          "$schema": {
            "examples": [
              "https://static.modelcontextprotocol.io/schemas/2025-12-11/server.schema.json"
            ],
            "minLength": 1,
            "title": "$Schema",
            "type": "string"
          },
          "_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "icons": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Icons",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$",
            "title": "Name",
            "type": "string"
          },
          "packages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Packages",
            "type": "array"
          },
          "remotes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Remotes",
            "type": "array"
          },
          "repository": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository"
          },
          "title": {
            "default": "",
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          },
          "websiteUrl": {
            "default": "",
            "maxLength": 2048,
            "title": "Websiteurl",
            "type": "string"
          }
        },
        "required": [
          "$schema",
          "name",
          "description",
          "version"
        ],
        "title": "RegistryServerVersionUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Wardn Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/mcp/servers": {
      "post": {
        "operationId": "admin_mcp_servers_create_version",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryServerVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerVersionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Create Mcp Server Version",
        "tags": [
          "admin-mcp"
        ]
      }
    },
    "/api/v1/admin/mcp/servers/{server_name}/versions/{version}": {
      "delete": {
        "operationId": "admin_mcp_servers_delete_version",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Delete Mcp Server Version",
        "tags": [
          "admin-mcp"
        ]
      },
      "put": {
        "operationId": "admin_mcp_servers_update_version",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryServerVersionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerVersionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Update Mcp Server Version",
        "tags": [
          "admin-mcp"
        ]
      }
    },
    "/api/v1/admin/mcp/servers/{server_name}/versions/{version}/latest": {
      "post": {
        "operationId": "admin_mcp_servers_set_latest_version",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerVersionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Set Latest Mcp Server Version",
        "tags": [
          "admin-mcp"
        ]
      }
    },
...
    "/api/v1/mcp/servers": {
      "get": {
        "operationId": "mcp_servers_list",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "updated_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated Since"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "latest",
              "title": "Version"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "support_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Support Level"
            }
          },
          {
            "in": "query",
            "name": "partner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Partner"
            }
          },
          {
            "in": "query",
            "name": "registry_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registry Type"
            }
          },
          {
            "in": "query",
            "name": "transport_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Transport Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Mcp Servers",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{server_name}": {
      "get": {
        "operationId": "mcp_servers_get",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mcp Server",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{server_name}/versions": {
      "get": {
        "operationId": "mcp_servers_list_versions",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerVersionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Mcp Server Versions",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{server_name}/versions/{version}": {
      "get": {
        "operationId": "mcp_servers_get_version",
        "parameters": [
          {
            "in": "path",
            "name": "server_name",
            "required": true,
            "schema": {
              "title": "Server Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryServerVersionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mcp Server Version",
        "tags": [
          "mcp"
        ]
      }
    },
...
  }
}